move most of index.js into App.js
